More warning squashings; sorry it's taking longer than I thought. Sometimes they make you think...

+++ src/gfx/nav/navgetxmldata.cpp
@@ -46,7 +46,7 @@ string retrievedata( string data, string type )
                     //***********************************
                     if (possibletype == type) {
                         //cout << possibletype;
-                        int    returncounter = counter;
+                        size_t returncounter = counter;
                         string returnstring  = "";
                         returncounter += 1;                             //pass the =
                         while (data[returncounter] == ' ')
